+++ config/bugsnag.php
@@ -123,17 +123,47 @@ return [
 
     /*
     |--------------------------------------------------------------------------
+    | Project Root Regex
+    |--------------------------------------------------------------------------
+    |
+    | Bugsnag marks stacktrace lines as in-project if they come from files
+    | inside your “project root”. You can set this here.
+    |
+    | This option allows you to set it as a regular expression and will take
+    | precedence over "project_root" if both are defined.
+    |
+    */
+
+    'project_root_regex' => env('BUGSNAG_PROJECT_ROOT_REGEX'),
+
+    /*
+    |--------------------------------------------------------------------------
     | Strip Path
     |--------------------------------------------------------------------------
     |
-    | You can set a strip path to have it also trimmed from the start of any
-    | filepath in your stacktraces.
+    | The strip path is a path to be trimmed from the start of any filepaths in
+    | your stacktraces.
     |
     | If this is not set, we will automatically try to detect it.
     |
     */
 
     'strip_path' => env('BUGSNAG_STRIP_PATH'),
+
+    /*
+    |--------------------------------------------------------------------------
+    | Strip Path Regex
+    |--------------------------------------------------------------------------
+    |
+    | The strip path is a path to be trimmed from the start of any filepaths in
+    | your stacktraces.
+    |
+    | This option allows you to set it as a regular expression and will take
+    | precedence over "strip_path" if both are defined.
+    |
+    */
+
+    'strip_path_regex' => env('BUGSNAG_STRIP_PATH_REGEX'),
 
     /*
     |--------------------------------------------------------------------------