+++ tests/ServiceProviderTest.php
@@ -6,6 +6,7 @@ use Bugsnag\BugsnagLaravel\LaravelLogger;
 use Bugsnag\BugsnagLaravel\MultiLogger;
 use Bugsnag\BugsnagLaravel\Queue\Tracker;
 use Bugsnag\Client;
+use Bugsnag\Configuration;
 use Bugsnag\PsrLogger\BugsnagLogger;
 use Bugsnag\PsrLogger\MultiLogger as BaseMultiLogger;
 use GrahamCampbell\TestBenchCore\ServiceProviderTrait;
@@ -16,6 +17,13 @@ use Mockery;
 class ServiceProviderTest extends AbstractTestCase
 {
     use ServiceProviderTrait;
+
+    /**
+     * An Application instance provided by the parent test case.
+     *
+     * @var Application
+     */
+    protected $app;
 
     public function testClientIsInjectable()
     {
@@ -35,6 +43,233 @@ class ServiceProviderTest extends AbstractTestCase
     public function testBugsnagLoggerIsInjectable()
     {
         $this->assertIsInjectable(interface_exists(Log::class) ? LaravelLogger::class : BugsnagLogger::class);
+    }
+
+    /**
+     * Ensure the project root and strip path are both set with sensible defaults
+     * when no explicit configuration is provided.
+     *
+     * @return void
+     */
+    public function testProjectRootAndStripPathAreInferredWhenNoSpecificConfigurationIsGiven()
+    {
+        $client = $this->app->make(Client::class);
+
+        $this->assertInstanceOf(Client::class, $client);
+
+        /** @var Client $client */
+        $config = $client->getConfig();
+
+        $projectRootRegex = $this->getProperty($config, 'projectRootRegex');
+        $stripPathRegex = $this->getProperty($config, 'stripPathRegex');
+
+        $expectedProjectRootRegex = $this->pathToRegex($this->app->path());
+        $expectedStripPathRegex = $this->pathToRegex($this->app->basePath());
+
+        $this->assertSame(
+            $expectedStripPathRegex,
+            $stripPathRegex,
+            "Expected to set a sensible default for the 'stripPathRegex'"
+        );
+        $this->assertSame(
+            $expectedProjectRootRegex,
+            $projectRootRegex,
+            "Expected to set a sensible default for the 'projectRootRegex'"
+        );
+    }
+
+    /**
+     * @param string|null $projectRoot
+     * @param string|null $stripPath
+     * @param string|null $projectRootRegex
+     * @param string|null $stripPathRegex
+     * @param string|null $expectedProjectRootRegex
+     * @param string      $expectedStripPathRegex
+     *
+     * @return void
+     *
+     * @dataProvider projectRootAndStripPathProvider
+     */
+    public function testProjectRootAndStripPathAreSetCorrectly(
+        $projectRoot,
+        $stripPath,
+        $projectRootRegex,
+        $stripPathRegex,
+        $expectedProjectRootRegex,
+        $expectedStripPathRegex
+    ) {
+        /** @var \Illuminate\Config\Repository $laravelConfig */
+        $laravelConfig = $this->app->config;
+        $bugsnagConfig = $laravelConfig->get('bugsnag');
+
+        $this->assertNull(
+            $bugsnagConfig['project_root'],
+            "Expected the default configuration value for 'project_root' to be null"
+        );
+
+        $this->assertNull(
+            $bugsnagConfig['strip_path'],
+            "Expected the default configuration value for 'strip_path' to be null"
+        );
+
+        $this->assertNull(
+            $bugsnagConfig['project_root_regex'],
+            "Expected the default configuration value for 'project_root_regex' to be null"
+        );
+
+        $this->assertNull(
+            $bugsnagConfig['strip_path_regex'],
+            "Expected the default configuration value for 'strip_path_regex' to be null"
+        );
+
+        $bugsnagConfig['project_root'] = $projectRoot;
+        $bugsnagConfig['strip_path'] = $stripPath;
+        $bugsnagConfig['project_root_regex'] = $projectRootRegex;
+        $bugsnagConfig['strip_path_regex'] = $stripPathRegex;
+
+        $laravelConfig->set('bugsnag', $bugsnagConfig);
+
+        $client = $this->app->make(Client::class);
+
+        $this->assertInstanceOf(Client::class, $client);
+
+        $appRoot = $this->app->path();
+
+        /** @var Client $client */
+        $config = $client->getConfig();
+
+        $projectRootRegex = $this->getProperty($config, 'projectRootRegex');
+        $stripPathRegex = $this->getProperty($config, 'stripPathRegex');
+
+        $this->assertSame(
+            $expectedProjectRootRegex,
+            $projectRootRegex,
+            "Expected the 'projectRootRegex' to match the string provided in Bugsnag configuration"
+        );
+
+        $this->assertSame(
+            $expectedStripPathRegex,
+            $stripPathRegex,
+            "Expected the 'stripPathRegex' to match the string provided in Bugsnag configuration"
+        );
+    }
+
+    public function projectRootAndStripPathProvider()
+    {
+        return [
+            // If both strings are provided, both options should be set to the
+            // regex version of the given strings
+            'both strings provided' => [
+                'project_root' => '/example/project/root',
+                'strip_path' => '/example/strip/path',
+                'project_root_regex' => null,
+                'strip_path_regex' => null,
+                'expected_project_root_regex' => $this->pathToRegex('/example/project/root'),
+                'expected_strip_path_regex' => $this->pathToRegex('/example/strip/path'),
+            ],
+
+            // If both regexes are provided they should be set verbatim
+            'both regexes provided' => [
+                'project_root' => null,
+                'strip_path' => null,
+                'project_root_regex' => '/^example project root regex/',
+                'strip_path_regex' => '/^example strip path regex/',
+                'expected_project_root_regex' => '/^example project root regex/',
+                'expected_strip_path_regex' => '/^example strip path regex/',
+            ],
+
+            // If only the project root string is provided, the project root should
+            // be set to the regex version of the string and the strip path to
+            // the application base path
+            'only project root string provided' => [
+                'project_root' => '/example/project/root',
+                'strip_path' => null,
+                'project_root_regex' => null,
+                'strip_path_regex' => null,
+                'expected_project_root_regex' => $this->pathToRegex('/example/project/root'),
+                'expected_strip_path_regex' => $this->pathToRegex($this->getBasePath()),
+            ],
+
+            // If only the project root regex is provided, both values should be
+            // set to the same regex
+            'only project root regex provided' => [
+                'project_root' => null,
+                'strip_path' => null,
+                'project_root_regex' => '/^example project root regex/',
+                'strip_path_regex' => null,
+                'expected_project_root_regex' => '/^example project root regex/',
+                'expected_strip_path_regex' => $this->pathToRegex($this->getBasePath()),
+            ],
+
+            // If only the strip path string is provided, both values should be
+            // set — the stip path to the regex version of the string and the
+            // project root with "/app" appended
+            'only strip path string provided' => [
+                'project_root' => null,
+                'strip_path' => '/example/strip/path',
+                'project_root_regex' => null,
+                'strip_path_regex' => null,
+                'expected_project_root_regex' => $this->pathToRegex("{$this->getBasePath()}/app"),
+                'expected_strip_path_regex' => $this->pathToRegex('/example/strip/path'),
+            ],
+
+            // If only the strip path regex is provided, the strip path should be
+            // set verbatim and the project root should be set to the application
+            // path (i.e. "/base/path/app")
+            'only strip path regex provided' => [
+                'project_root' => null,
+                'strip_path' => null,
+                'project_root_regex' => null,
+                'strip_path_regex' => '/^example strip path regex/',
+                'expected_project_root_regex' => $this->pathToRegex("{$this->getBasePath()}/app"),
+                'expected_strip_path_regex' => '/^example strip path regex/',
+            ],
+
+            // If the regexes are provided and either string value is too then
+            // the regexes should take precedence and the string value ignored
+            'project root string and both regexes provided' => [
+                'project_root' => $this->pathToRegex('/example/project/root'),
+                'strip_path' => null,
+                'project_root_regex' => '/^example project root regex/',
+                'strip_path_regex' => '/^example strip path regex/',
+                'expected_project_root_regex' => '/^example project root regex/',
+                'expected_strip_path_regex' => '/^example strip path regex/',
+            ],
+
+            // If the regexes are provided and either string value is too then
+            // the regexes should take precedence and the string value ignored
+            'strip path string and both regexes provided' => [
+                'project_root' => null,
+                'strip_path' => $this->pathToRegex('/example/strip/path'),
+                'project_root_regex' => '/^example project root regex/',
+                'strip_path_regex' => '/^example strip path regex/',
+                'expected_project_root_regex' => '/^example project root regex/',
+                'expected_strip_path_regex' => '/^example strip path regex/',
+            ],
+
+            // If all four options are provided then the regexes should take
+            // precedence and the string values ignored
+            'all options provided' => [
+                'project_root' => $this->pathToRegex('/example/project/root'),
+                'strip_path' => $this->pathToRegex('/example/strip/path'),
+                'project_root_regex' => '/^example project root regex/',
+                'strip_path_regex' => '/^example strip path regex/',
+                'expected_project_root_regex' => '/^example project root regex/',
+                'expected_strip_path_regex' => '/^example strip path regex/',
+            ],
+        ];
+    }
+
+    /**
+     * Convert a file path to a regex that matches the path and any sub paths.
+     *
+     * @param string $path
+     *
+     * @return string
+     */
+    private function pathToRegex($path)
+    {
+        return sprintf('/^%s[\\/]?/i', preg_quote($path, '/'));
     }
 
     public function testCorrectLoggerClassesReturned()
@@ -59,7 +294,6 @@ class ServiceProviderTest extends AbstractTestCase
                 function ($closure) use ($loggerClass) {
                     if (is_callable($closure)) {
                         $internalApp = Mockery::mock(Application::class);
-                        $config = Mockery::mock();
                         $internalApp->shouldReceive('offsetGet')->with('config')->andReturn($internalApp);
                         $internalApp->shouldReceive('get')->with('bugsnag')->andReturn([
                             'logger_notify_level' => 'error',